requirements.txt
pydantic-ai
requests
python-dotenv